# api/src/lib/openapi/base.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenZenith API",
    "version": "0.9.2",
    "description": "Free, fast, global elevation and geospatial API. Query any point on Earth for elevation data from NASA SRTM 30m, track flights, monitor weather, explore OpenStreetMap, and more. No API key required.",
    "contact": {
      "name": "OpenZenith",
      "url": "https://github.com/aliasfoxkde/OpenZenith"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment (derived from request)"
    }
  ],
  "paths": {
...
    "/api/overpass": {
      "post": {
        "summary": "Query OpenStreetMap (Overpass)",
        "description": "Proxies Overpass API queries against OpenStreetMap data. Supports full Overpass QL syntax. Max query length: 10,000 characters. 30-second timeout. 60-second cache.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "Overpass QL query string",
                    "maxLength": 10000
                  }
                }
              },
              "example": {
                "query": "[out:json];node(48.85,2.35,48.86,2.36)[amenity=cafe];out 3;"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Overpass API result (JSON)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Overpass API response with elements array",
                  "properties": {
                    "version": {
                      "type": "number",
                      "example": 0.6
                    },
                    "generator": {
                      "type": "string",
                      "example": "Overpass API"
                    },
                    "elements": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query"
          },
          "502": {
            "description": "Overpass API error or timeout"
          }
        },
        "tags": ["Data"]
      }
    },
...
    "/api/waterways": {
      "get": {
        "summary": "Get water features in a bounding box",
        "description": "Returns rivers, lakes, and other water features within a bounding box from OpenStreetMap via the Overpass API. Results are returned as a GeoJSON FeatureCollection. Cached for 7 days.",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Bounding box as minLon,minLat,maxLon,maxLat",
            "example": "-74.1,40.6,-73.9,40.8"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["rivers", "lakes", "all"]
            },
            "description": "Feature type to return (default: all)",
            "example": "rivers"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500
            },
            "description": "Maximum features (default: 100)",
            "example": 100
          }
        ],
        "responses": {
          "200": {
            "description": "GeoJSON FeatureCollection of water features",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "GeoJSON FeatureCollection",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "FeatureCollection"
                    },
                    "features": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid bbox"
          },
          "502": {
            "description": "Overpass API unavailable"
          }
        },
        "tags": ["Hydrography"]
      }
    },
...
  },
...
}